### 2021-09-03 #3692 @whedon
[PRE REVIEW]: UnlockNN: Uncertainty quantification for neural network models of chemical systems
**Submitting author:** @a-ws-m (<a href="http://orcid.org/0000-0001-7525-1419">Alexander Moriarty</a>)
**Repository:** <a href="https://github.com/a-ws-m/unlockNN/" target ="_blank">https://github.com/a-ws-m/unlockNN/</a>
**Version:** v2.0.2
