### 2021-09-03 #3692 @kyleniemeyer
Hi  @a-ws-m, thanks for your interest in JOSS! Unfortunately due to low editor availability in this area, I have to put your submission on our waitlist. We'll get it assigned and moving forward as soon as someone opens up to handle this.

In the meantime, could you take a look at that DOI issue in the article? Any reviewer suggestions could share would be much appreciated as well.
