### 2021-09-04 #3692 @a-ws-m
Thanks @kyleniemeyer, I've fixed that DOI issue. One potential reviewer could be JaGeo (though she doesn't appear on the list; I'll have a look through that, too).
