+++ README.md
@@ -2,6 +2,7 @@
 
 [![Documentation Status](https://readthedocs.org/projects/unlockgnn/badge/?version=latest)](https://unlockgnn.readthedocs.io/en/latest/?badge=latest)
 ![PyPi](https://img.shields.io/pypi/v/unlockNN)
+[![status](https://joss.theoj.org/papers/b00df538a159c4b6816ec24d4d1716fb/status.svg)](https://joss.theoj.org/papers/b00df538a159c4b6816ec24d4d1716fb)
 [![made-with-python](https://img.shields.io/badge/Made%20with-Python-1f425f.svg)](https://www.python.org/)
 [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
 [![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
