### 2021-09-06 #3692 @TahiriNadia
Thank you @osorensen. I am interested in reviewing the paper. As this is my first for this journal, would it be possible to give me the link for the directions (containing the evaluation grid)?
