### 2021-09-07 #3692 @pmeier
@osorensen Depends on what you are looking for. I have no experience with _variational Gaussian processes_ (VGPs), which seem to play a major part in the package. Willing to look into them though.
