### 2021-09-07 #3692 @osorensen
@pmeier, thanks for responding so quickly. From your GitHub profile, you seem highly competent in software for machine learning, and I hence think you would be a very good reviewer for this submission. In the typical JOSS papers the underlying methodology should be described elsewhere and referenced, so the reviewers don't necessarily need to be experts on all those aspects.

If that sounds fine with you I'll add you as a reviewer.
