### 2021-09-07 #3692 @Het-Shah
@osorensen Thank you for the request to review. I agree with @pmeier, I do not have experience with VGPs. I am willing to have a look through the work if a reviewer is still needed.
