### 2021-09-07 #3692 @osorensen
> @osorensen Can you tell me how many days we have to revise the article? please.

@TahiriNadia, I hope that you can starting with the review within 2-3 weeks, but we try to be flexible in these special times. It is very hard to say how long the review process will last. In JOSS reviews we typically let the submitting authors improve the software based on the reviewers' feedback, so it may be that you first do an initial review, and then come back in a month or so to evaluate whether the software has been approved according to your suggestions.
