### 2021-09-07 #3700 @whedon
[REVIEW]: UnlockNN: Uncertainty quantification for neural network models of chemical systems
**Submitting author:** @a-ws-m (<a href="http://orcid.org/0000-0001-7525-1419">Alexander Moriarty</a>)
**Repository:** https://github.com/a-ws-m/unlockNN/
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.2

- [x] #3793
- [x] #3796
- [x] #3798
- [x] #3799

- [x] #3797
