### 2021-09-27 #3700 @pmeier
@osorensen I've checked all the basic things and so far everything looks good. The paper is well written. The only thing left for me is to check its functionality and the documentation thereof. This might take some time, since I'm neither familiar with TensorFlow / keras nor have any domain knowledge. Regarding the latter, I fail to see how I should provide an opinion if the literature references are complete or not.
