### 2021-09-27 #3700 @osorensen
Thanks for replying, @pmeier! Regarding the literature references, I assume you're referring to this part:

> Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?

I suggest you can interpret this as asking whether papers/datasets/software mentioned inside the paper are properly cited, and whether the citation syntax is correct. So if you don't have an overview of this particular application area, you may still check off the box if the citations are done formally correct.
