### 2021-09-27 #3700 @TahiriNadia
Sorry, My ckeck list is not clikeable (see below).
![image](https://user-images.githubusercontent.com/19578926/134919516-10e5c060-f8f7-401e-beb1-4cf6bdcf645d.png)
