### 2021-09-27 a-ws-m/unlockNN#19 @Het-Shah
JOSS Review
Hi @a-ws-m! This will be an issue for my review for your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/3700).

Pros -
- The API is straightforward for the end-user to use.

Cons -
- The library seems to have ~1000 lines of actual code (1165). This begs the question about the substantial scholarly effort.
- The paper lacks details about the library itself. It is unclear as to how the package is structured and what are the key functionalities.
- It would be advantageous if you could make some basic ablations and provide their results. I am unaware of the datasets, but you can try on the smallest dataset and provide results.

Feel free to comment on these. I would be more than happy to discuss the same.
