### 2021-09-28 #3700 @osorensen
@pmeier, regarding what you wrote yesterday:

>  The only thing left for me is to check its functionality and the documentation thereof.

A general comment, which may or may not apply to this particular case: If you find it hard to test functionality and run the examples, it may well be that other users would experience the same. You might consider whether this is a point at which the software could be improved, and if so, feel free to add a suggestion here or create an issue in the source repository.
