### 2021-10-03 a-ws-m/unlockNN#20 @pmeier
RTD slug still reads `unlockgnn`
https://unlockgnn.readthedocs.io/

If I'm not mistaken, this was changed in f1c113e13628a70656557338ba318a33027daa0d.
