### 2021-10-03 a-ws-m/unlockNN#21 @pmeier
What is `issue #152`?
All test modules under `tests/` set the `CUDA_DEVICE_ORDER` environment variable with the hint to look at `issue #152`:

https://github.com/a-ws-m/unlockNN/blob/128b192770925220e24761173c9c373d90088868/tests/test_download.py#L5-L6

This issue does not exist in this repository so the hint should probably a proper link.
