### 2021-10-03 a-ws-m/unlockNN#22 @pmeier
`test_download_model` doesn't test anything?
The test simply calls `load_pretrained` twice, but doesn't do anything with the results:

https://github.com/a-ws-m/unlockNN/blob/128b192770925220e24761173c9c373d90088868/tests/test_download.py#L11-L17
