### 2021-10-03 a-ws-m/unlockNN#23 @pmeier
`test_load_data` checks for floating point equality
https://github.com/a-ws-m/unlockNN/blob/128b192770925220e24761173c9c373d90088868/tests/test_download.py#L22-L23

In general, we should never check floating point values for equality. Especially the last six digits look insignificant where the `1` is most likely an artifact of the imperfect floating point representation. Maybe [`pytest.approx`](https://docs.pytest.org/en/6.2.x/reference.html#pytest.approx) would be a better fit here?
