### 2021-10-03 a-ws-m/unlockNN#24 @pmeier
Attribution of `datadir` fixture is insufficient
The docstring currently only links to an StackOverflow answer:

https://github.com/a-ws-m/unlockNN/blob/128b192770925220e24761173c9c373d90088868/tests/test_model.py#L34

StackOverflow answers are published under the [CC BY-SA 4.0](https://creativecommons.org/licenses/by-sa/4.0/) license, which requires more. Especially the ShareAlike part might be problematic, since `unlockNN` is distributed under the MIT license.
