### 2021-10-03 a-ws-m/unlockNN#25 @pmeier
`test_meg_prob` seems to test multiple things at once and is hard to follow
Maybe split it into multiple smaller ones? Rule of thumb should be to only have a single `assert` per test.
