### 2021-10-03 a-ws-m/unlockNN#26 @pmeier
Code coverage is fairly low
Although the coverage is 78%, the tests cover very little apart from the regular cases. A lot of `if` branches especially in error cases are not covered. Additionally, the following functions and modules are not covered at all:

- `unlocknn.metrics`
- `unlocknn.megnet_utils`
    - `.default_megnet_config()`
    - `.scale_targets()`
