### 2021-10-03 a-ws-m/unlockNN#29 @pmeier
Contributing guidelines do not explain how to use the required tools
The guidelines mention `black` and `pydocstyle`, but no instructions how to use them. Additionally

> utilities for building documentation, enabling the testing suite and publishing to PyPi

is fairly vague and one has to dig through the CI workflows for example to find how the tests should be run.
