### 2021-10-03 a-ws-m/unlockNN#30 @pmeier
Running the test suite emits 27 warnings
See for example [this CI run](https://github.com/a-ws-m/unlockNN/runs/3513741203). Most of them seem to be because deprecated behavior is used or functionalities are not properly invoked.
