### 2021-10-03 a-ws-m/unlockNN#31 @pmeier
malfunctional metrics?
The documentation of `MEGNetProbModel` warns in two locations, that the `metrics` are "malfunctional". If that is the case, why provide the ability to use malfunctional ones and not the working ones?

On a side note, pointing the users the `unlocknn.metrics` is a little peculiar in the light of #26.
