### 2021-10-03 a-ws-m/unlockNN#32 @pmeier
Documentation for `AmpAndLengthScale` mentions internal attributes 
If `amplitude` and `_length_scale` should be public, we should remove the leading underscore. Otherwise they should be removed from the documentation.
