### 2021-10-03 a-ws-m/unlockNN#33 @pmeier
The uncertainty quantifier-specific metrics could use some links
Given that the documentation is minimal (which is not a bad thing)

https://github.com/a-ws-m/unlockNN/blob/128b192770925220e24761173c9c373d90088868/unlocknn/metrics.py#L19

a link to Wikipedia or equivalent would probably help a user to better understand the function.
