### 2021-10-03 #3700 @pmeier
I've added a bunch of issues about details for my review:

## Automated tests and contributing

- a-ws-m/unlockNN#21
- a-ws-m/unlockNN#22
- a-ws-m/unlockNN#23
- a-ws-m/unlockNN#24
- a-ws-m/unlockNN#25
- a-ws-m/unlockNN#26
- a-ws-m/unlockNN#29
- a-ws-m/unlockNN#30

## Documentation

- a-ws-m/unlockNN#20
- a-ws-m/unlockNN#31
- a-ws-m/unlockNN#32
- a-ws-m/unlockNN#33

---

In general though, I currently agree with the points @Het-Shah brought up in a-ws-m/unlockNN#19. Regarding the substantial scholarly effort, in the 1165 LoC a large amount is actually periphery such as downloading example weights or data or specifically tailored to the MEGNet models. If one removes that, `unlockNN` seems to "only" provide functionality to replace the last layer of a model with one that can also quantify the uncertainty of the prediction as well as metrics to evaluate such a model. I quoted _only_ above, since I'm maybe misreading this. I'm open to a discussion about this to convince me that I'm wrong.
