### 2021-10-03 #3700 @osorensen
Regarding the issue of substantial scholarly effort, as raised by @pmeier and @Het-Shah, the [guidelines](https://joss.readthedocs.io/en/latest/submitting.html) say the following:

> JOSS publishes articles about software that represent substantial scholarly effort on the part of the authors. Your software should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler).
  >
 > As a rule of thumb, JOSS’ minimum allowable contribution should represent not less than three months of work for an
 individual. Some factors that may be considered by editors and reviewers when judging effort include:
  >
  > - Age of software (is this a well-established software project) / length of commit history.
  > - Number of commits.
  > - Number of authors.
>   - Total lines of code (LOC). Submissions under 1000 LOC will usually be flagged, those under 300 LOC will be desk rejected.
 > - Whether the software has already been cited in academic papers.
 > - Whether the software is sufficiently useful that it is likely to be cited by your peer group.
 > - In addition, JOSS requires that software should be feature-complete (i.e. no half-baked solutions) and designed for maintainable extension (not one-off modifications of existing tools). “Minor utility” packages, including “thin” API clients, and single-function packages are not acceptable.

Hence, it is not only an issue of number of lines of code. @pmeier, @Het-Shah, and @a-ws-m, it would be great to hear your thoughts on this. Particularly if more of these points would be satisified if the issues raised by the reviewers are addressed.
