### 2021-10-03 #3700 @TahiriNadia
> Thanks for noting this, @pmeier.
>
> @TahiriNadia, I'm sorry about this confusing aspect - you're not the first reviewer who has done this. Could you please open you issues in the source repository instead as @pmeier suggests?

Thanks for the comment @osorensen, as I mentioned I am new to this article evaluation protocol. But I will look at the @pmeier comments as you mention. Thank you!
