### 2021-10-04 #3700 @pmeier
@osorensen my take on https://github.com/openjournals/joss-reviews/issues/3700#issuecomment-932977470

- :green_circle: **Age of software (is this a well-established software project) / length of commit history**: first commit at July 2020
- :green_circle: **Number of commits**: ~500
- :red_circle: **Number of authors**: `git shortlog -s` reveals two authors, which I think are the same: "Alex Moriarty" and "a-ws-m". This shouldn't be blocking.
- :yellow_circle: **Total lines of code (LOC). Submissions under 1000 LOC will usually be flagged, those under 300 LOC will be desk rejected** see my comments in https://github.com/openjournals/joss-reviews/issues/3700#issuecomment-932956088
- :red_circle: **Whether the software has already been cited in academic papers**: a quick search turned up empty, but this also hard without DOI. This shouldn't be blocking.
- :yellow_circle: **Whether the software is sufficiently useful that it is likely to be cited by your peer group**: hard to judge that without domain knowledge. Let's hope the discussion makes this more clear.
- :green_circle: **In addition, JOSS requires that software should be feature-complete (i.e. no half-baked solutions) and designed for**: package seems complete and flexible enough to not be a one off script
