+++ tests/test_model.py
@@ -31,8 +31,6 @@ def datadir(tmpdir, request):
     module and, if available, moving all contents to a temporary directory so
     tests can use them freely.
 
-    Source: https://stackoverflow.com/a/29631801/
-
     """
     filename = request.module.__file__
     test_dir, _ = os.path.splitext(filename)
