### 2021-10-14 7c45497
Make save_struct_data public (#27)
