+++ unlocknn/download.py
@@ -151,7 +151,7 @@ def _load_struct_data(fname: PathLike) -> pd.DataFrame:
     return serial_df
 
 
-def _save_struct_data(df: pd.DataFrame, fname: PathLike):
+def save_struct_data(df: pd.DataFrame, fname: PathLike):
     """Save data containing ``Structure``s to a file.
 
     Serializes the "structure" column to a string.
