### 2021-10-14 80644f6
Actually test downloading (#22)
