### 2021-10-14 91a3651
Add pytest-cov to tox deps
