+++ tox.ini
@@ -13,6 +13,7 @@ deps =
     tensorflow-probability>=0.10.1
     pymatgen<=2021.2.8
     pytest
+    pytest-cov
     ; pytest-xdist
     pandas
     pyarrow>=1.0.1
