### 2021-10-14 20a5957
Use pytest.approx for floating points (#23)
