### 2021-10-14 c2cfbf1
Remove reference to tensorflow issue
