+++ tests/test_kernels.py
@@ -1,7 +1,7 @@
 """Test the `kernel_layers` package."""
 import os
 
-os.environ["CUDA_DEVICE_ORDER"] = "PCI_BUS_ID"  # see issue #152
+os.environ["CUDA_DEVICE_ORDER"] = "PCI_BUS_ID"
 os.environ["CUDA_VISIBLE_DEVICES"] = ""
 
 from pathlib import Path
