### 2021-10-14 1112af1
Correct pytest.approx syntax (#23)
