### 2021-10-14 37ea3bf
Test reload functionality in load_data
