### 2021-10-15 fafe25a
[WIP] Metrics testing suite
