### 2021-10-15 a-ws-m/unlockNN#30 @pmeier
> - This is expected behaviour, and `cdf` is not needed.

Expected as in "I want to test that the warning is raised" or rather "I know this raises a warning, but the call is actually correct"? If it's the first, I recommend [`pytest.warns`](7c45497616195e93e47737b41aef11e33b254d2e) to assert that this is happening. In the latter case, it might be beneficial to silence the warning locally for example with [`warnings.catch_warnings`](https://docs.python.org/3/library/warnings.html#warnings.catch_warnings). Otherwise, as a user I would always be in doubt if I made a mistake and the warning relates to my code.

The last argument also applies to the other points you raised. If a user might hit these warnings, you should probably silence them locally. If they are only hit in the test suite you can also filter these directly by putting

```ini
addopts =
  -Wignore::UserWarning::Converting sparse ...
```

in your `pytest` configuration.
