### 2021-10-15 a-ws-m/unlockNN#33 @pmeier
Could you build the [`joss-review` branch](https://github.com/a-ws-m/unlockNN/tree/joss-review) on RTD so I don't have to read the raw RST? :innocent:
