### 2021-10-21 5683f30
Add save_struct_data to docs (#27)
