+++ unlocknn/download.py
@@ -152,10 +152,10 @@ def _load_struct_data(fname: PathLike) -> pd.DataFrame:
 
 
 def save_struct_data(df: pd.DataFrame, fname: PathLike):
-    """Save data containing ``Structure``s to a file.
+    """Save data containing ``Structure`` s to a file.
 
     Serializes the "structure" column to a string.
-    The converse of :func:`load_struct_data`.
+    The converse of :func:`_load_struct_data`.
 
     Args:
         df: The :class:`pd.DataFrame` to serialize.
