+++ tests/test_model.py
@@ -22,9 +22,10 @@ np.random.seed(123)
 python_random.seed(123)
 tf.random.set_seed(123)
 
+SplitData = Tuple[Tuple[list, list], Tuple[list, list]]
 
 @pytest.fixture
-def datadir(tmpdir, request):
+def datadir(tmpdir, request) -> Path:
     """Access data directory.
 
     Fixture responsible for searching a folder with the same name of test
@@ -46,13 +47,31 @@ def datadir(tmpdir, request):
 def weights_equal(weights_a: List[np.ndarray], weights_b: List[np.ndarray]) -> bool:
     """Check equality between weights."""
     return all(
-        (weight1 == weight2).all() for weight1, weight2 in zip(weights_a, weights_b)
+        weight1 == pytest.approx(weight2, rel=1e-6) for weight1, weight2 in zip(weights_a, weights_b)
     )
+
+def load_df_head(fname: Path, num_entries: int=100) -> pd.DataFrame:
+    """Load first entries of a pandas DataFrame in a backwards-compatible way.
+    
+    Args:
+        fname: The pickle file to open.
+        num_entries: How many values to read.
+    
+    """
+    try:
+        return pd.read_pickle(fname)[:num_entries]
+    except ValueError:
+        # Older python version
+        import pickle5 as pkl
+
+        with fname.open("rb") as f:
+            return pkl.load(f)[:num_entries]
+
 
 
 def train_test_split(
     structures: list, targets: list, train_frac: float = 0.8
-) -> Tuple[Tuple[list, list], Tuple[list, list]]:
+) -> SplitData:
     """Split structures and targets into training and testing subsets."""
     num_train = floor(len(structures) * train_frac)
     return (
@@ -60,56 +79,127 @@ def train_test_split(
         (structures[num_train:], targets[num_train:]),
     )
 
-
-def test_sample_init(datadir: Path):
-    """Test the SampleInitializer."""
-    megnet_e_form_model = MEGNetModel.from_file(str(datadir / "formation_energy.hdf5"))
+@pytest.fixture
+def split_data(datadir: Path) -> SplitData:
+    """Get some example data, split into training and test subsets.
+    
+    Returns:
+        train_structs
+        
+        train_targets
+        
+        test_structs
+        
+        test_targets
+    
+    """
     binary_dir = datadir / "mp_binary_on_hull.pkl"
-
-    try:
-        binary_df = pd.read_pickle(binary_dir)[:100]
-    except ValueError:
-        # Older python version
-        import pickle5 as pkl
-
-        with binary_dir.open("rb") as f:
-            binary_df = pkl.load(f)
-
+    binary_df = load_df_head(binary_dir)
     structures = binary_df["structure"].tolist()
     formation_energies = binary_df["formation_energy_per_atom"].tolist()
 
-    (train_structs, _), (_, _) = train_test_split(structures, formation_energies)
+    return train_test_split(
+        structures, formation_energies
+    )
+
+
+def test_sample_init(datadir: Path, split_data: SplitData):
+    """Test the SampleInitializer."""
+    megnet_e_form_model = MEGNetModel.from_file(str(datadir / "formation_energy.hdf5"))
+    (train_structs, _), (_, _) = split_data
+
     initializer = SampleInitializer(train_structs, megnet_e_form_model, batch_size=32)
     MEGNetProbModel(megnet_e_form_model, 10, index_initializer=initializer)
     # If this works without any errors, we're doing OK
 
 
 @pytest.mark.parametrize("use_norm", [True, False])
-def test_meg_prob(tmp_path: Path, datadir: Path, use_norm: bool):
[…]
