### 2021-10-21 de19247
Add metrics tests (#26)
