### 2021-10-22 e22576e
Don't cover backwards compatible checks (use tox)
