### 2021-10-22 46b0911
Add a test for custom kernels
