### 2021-10-22 4298d05
Clarify amplitude and length_scale (#32)
