+++ tests/test_kernels.py
@@ -41,14 +41,22 @@ def test_custom_kernel_reloading(tmp_path: Path):
 @pytest.mark.parametrize("kernel_type", [RBFKernelFn, MaternOneHalfFn])
 def test_reload(tmp_path: Path, kernel_type: Type[AmpAndLengthScaleFn]):
     """Test saving and reloading a builtin kernel to/from disk."""
-    # Example layer weights: [_amplitude, _length_scale]
+    # Example layer weights: [_amplitude_basis, _length_scale_basis]
     example_weights = [np.array(2.0), np.array(3.0)]
     save_dir = tmp_path / "kernel"
     orig_kernel = kernel_type()
     orig_kernel.set_weights(example_weights)
     orig_kernel.save(save_dir)
 
-    loaded_kernel = load_kernel(save_dir)
+    orig_amplitude = orig_kernel.amplitude
+    orig_length_scale = orig_kernel.length_scale
+
+    loaded_kernel: AmpAndLengthScaleFn = load_kernel(save_dir)
     assert isinstance(loaded_kernel, kernel_type)
     loaded_weights = loaded_kernel.get_weights()
     assert loaded_weights == example_weights
+
+    loaded_amp = loaded_kernel.amplitude
+    loaded_ls = loaded_kernel.length_scale
+    assert loaded_amp == pytest.approx(orig_amplitude)
+    assert loaded_ls == pytest.approx(orig_length_scale)
