### 2021-10-22 57e793c
Correct TestKernel implementation to match docs
