### 2021-10-22 7072967
Add amplitude and ls to docs (#32)
