### 2021-10-22 ba6c627
[FIX] amplitude and ls property types (#32)
