### 2021-10-25 1dad84a
Warnings handling (#30)
