+++ pytest.ini
@@ -1,0 +1,10 @@
+[pytest]
+addopts = --cov=unlocknn
+testpaths =
+    tests
+filterwarnings =
+    ignore::DeprecationWarning
+    ignore::PendingDeprecationWarning
+    ignore:Unable to detect statically whether the number of index_points is 1:UserWarning
+    ignore:Converting sparse .* to a dense Tensor of unknown shape:UserWarning
+    ignore:Custom mask layers require a config and must override get_config:Warning
