### 2021-10-25 65ceb43
Add freezing/thawing tests (#26)
