### 2021-10-25 99579d4
MEGNet utils testing (#26)
