### 2021-10-25 5b68426
Add clarifying info on package structure (#19)
