+++ benchmarking/mp-2019/main.py
@@ -1,0 +1,168 @@
+"""Train and evaluate a MEGNetProbModel for formation energy prediction on the MP-2019 dataset.
+
+<https://figshare.com/articles/dataset/Graphs_of_Materials_Project_20190401/8097992>
+
+"""
+import datetime
+import json
+from argparse import ArgumentError, ArgumentParser
+from os import mkdir
+from pathlib import Path
+from typing import Any, Dict, Iterable, Optional
+
+import numpy as np
+import pandas as pd
+from megnet.models import MEGNetModel
+from megnet.utils.models import load_model
+from pymatgen.core.structure import Structure
+from tensorflow.python.keras.callbacks import TensorBoard
+from unlocknn import MEGNetProbModel
+from unlocknn.metrics import evaluate_uq_metrics, MSE
+
+HERE = Path(__file__).parent
+DATA_FILE = HERE / "mp.2019.04.01.json"
+MODELS_DIR = HERE / "models"
+LOGS_DIR = HERE / "logs"
+
+MEGNET_METRICS_LOG = HERE / "meg_metrics.log"
+PROB_MODEL_METRICS_LOG = HERE / "prob_metrics.json"
+
+TRAIN_RATIO: float = 0.9
+VAL_RATIO: float = 0.05
+
+for directory in [MODELS_DIR, LOGS_DIR]:
+    if not directory.exists():
+        mkdir(directory)
+
+def model_dir(num_inducing_points: int) -> Path:
+    """Get the models directory for the model with _n_ index points."""
+    return MODELS_DIR / f"mp2019_e_form_{num_inducing_points}"
+
+def log_dir(num_inducing_points: int) -> Path:
+    """Get the logs directory for the model with _n_ index points."""
+    log_dir = LOGS_DIR / f"mp2019_e_form_{num_inducing_points}"
+    if not log_dir.exists():
+        mkdir(log_dir)
+    return log_dir
+
+def get_tb_callback(num_inducing_points: int) -> TensorBoard:
+    """Get a configured TensorBoard callback."""
+    tensorboard_run_loc = log_dir(num_inducing_points) / str(
+        datetime.now().strftime("%Y.%m.%d.%H.%M.%S")
+    )
+
+    # These parameters avoid segmentation fault
+    return TensorBoard(
+        tensorboard_run_loc, write_graph=False, profile_batch=0
+    )
+
+def load_meg_model() -> MEGNetModel:
+    """Load the MEGNetModel."""
+    return load_model("Eform_MP_2019")
+
+def eval_meg_model(meg_model: MEGNetModel, structures: Iterable[Structure], targets: Iterable[float]):
+    """Evaluate the MSE of a given MEGNetModel."""
+    predictions = meg_model.predict_structures(structures)
+    return MSE(predictions, None, targets)
+
+def load_data() -> pd.DataFrame:
+    """Load the data from disk."""
+    with DATA_FILE.open("r") as f:
+        data = json.load(f)
+    
+    df = pd.DataFrame({
+        "structure": [Structure.from_str(entry["structure"]) for entry in data],
+        "e_form_per_atom": [entry["formation_energy_per_atom"] for entry in data],
+        "band_gap": [entry["band_gap"] for entry in data]
+    })
+    # Decorate DataFrame with set type
+    num_entries = len(df.index)
+    train_size = np.floor(TRAIN_RATIO * num_entries)
+    val_size = np.floor(VAL_RATIO * num_entries)
+
+    df["training_set"] = [i < train_size for i in range(num_entries)]
+    df["testing_set"] = [i >= train_size + val_size for i in range(num_entries)]
+    df["validation_set"] = [not (train or test) for train, test in zip(df["training_set"], df["testing_set"])]
+
+    # Sanity check
+    test_size = num_entries - train_size - val_size
[…]
