### 2021-10-26 75edfc7
Enable training and inference with graphs
