+++ unlocknn/model.py
@@ -5,7 +5,7 @@ from os import PathLike
 import warnings
 from abc import ABC, abstractmethod
 from pathlib import Path
-from typing import Any, Dict, Iterable, List, Optional, Tuple, Union
+from typing import Any, Dict, List, Optional, Tuple, Union
 
 try:
     from typing import Literal, get_args
@@ -23,7 +23,7 @@ from pymatgen.core import Structure
 from tensorflow.python.keras.utils import losses_utils
 
 from .kernel_layers import KernelLayer, RBFKernelFn, load_kernel
-from .megnet_utils import create_megnet_input, Targets
+from .megnet_utils import create_megnet_input, Targets, ModelInput
 
 tfd = tfp.distributions
 
@@ -534,10 +534,10 @@ class MEGNetProbModel(ProbNN):
 
     def train(
         self,
-        structs: List[Structure],
+        inputs: List[Structure],
         targets: Targets,
         epochs: int,
-        val_structs: Optional[List[Structure]] = None,
+        val_inputs: Optional[List[ModelInput]] = None,
         val_targets: Optional[Targets] = None,
         callbacks: List[tf.keras.callbacks.Callback] = [],
         use_default_ckpt_handler: bool = True,
@@ -549,7 +549,7 @@ class MEGNetProbModel(ProbNN):
         """Train the model.
 
         Args:
-            structs: A list of training crystal structures.
+            inputs: A list of training crystal structures or graphs.
             targets: A list of training target values.
             epochs: The number of training epochs.
             val_structs: A list of validation crystal structures.
@@ -568,17 +568,17 @@ class MEGNetProbModel(ProbNN):
         """
         # Convert structures to graphs for model input
         train_gen, train_graphs = create_megnet_input(
-            self.meg_model, structs, targets, batch_size, scrub_failed_structs
+            self.meg_model, inputs, targets, batch_size, scrub_failed_structs
         )
         steps_per_train = int(np.ceil(len(train_graphs) / batch_size))
 
         val_gen = None
         val_graphs = None
         steps_per_val = None
-        if val_structs is not None and val_targets is not None:
+        if val_inputs is not None and val_targets is not None:
             val_gen, val_graphs = create_megnet_input(
                 self.meg_model,
-                val_structs,
+                val_inputs,
                 val_targets,
                 batch_size,
                 scrub_failed_structs,
@@ -633,7 +633,7 @@ class MEGNetProbModel(ProbNN):
         )
 
     def predict(
-        self, input: Union[Structure, Iterable[Structure]], batch_size: int = 128
+        self, input: Union[ModelInput, List[ModelInput]], batch_size: int = 128
     ) -> Tuple[np.ndarray, np.ndarray]:
         """Predict target values and standard deviations for a given input.
 
@@ -665,12 +665,12 @@ class MEGNetProbModel(ProbNN):
         """
         self.update_pred_model()
 
-        if isinstance(input, Structure):
+        if not isinstance(input, list):
             # Just one to predict
             input = [input]
 
         inputs, graphs = create_megnet_input(
-            self.meg_model, structs=input, batch_size=batch_size, shuffle=False
+            self.meg_model, inputs=input, batch_size=batch_size, shuffle=False
         )
         num_atoms = [len(graph["atom"]) for graph in graphs]
 
