### 2021-10-26 98c27f3
Update kwargs to reflect possible graph input
