### 2021-10-26 cfe1c38
Update benchmark script to use graphs
