### 2021-10-26 c7be46d
Don't use band gap
