### 2021-10-26 5e3aff8
Fix case where no inducing points are set
