### 2021-10-27 a4e99b0
Add df debug statement
