### 2021-10-27 9cafceb
Training data debug
