### 2021-10-27 cf64388
Correct dataframe queries
