+++ benchmarking/mp-2019/main.py
@@ -162,6 +162,7 @@ def main():
 
     if cli_args["train"]:
         # * Train the probabilistic model
+        print(val_df["graph"].head())
         # Training time
         prob_model.train(train_df["graph"], train_df["e_form_per_atom"], cli_args["train"], val_df["graph"], val_df["e_form_per_atom"], callbacks=[get_tb_callback(NUM_INDUCING_POINTS)])
         prob_model.save(MODEL_DIR)
