### 2021-11-03 #3700 @osorensen
@Het-Shah, could you please check whether your points raised in [this issue](https://github.com/a-ws-m/unlockNN/issues/19) have now been properly addressed by the author, and if so, update your review checklist at the top of this issue?
