### 2021-11-03 #3700 @Het-Shah
@osorensen I believe @a-ws-m is still working on benchmarks. Once these experiments are done, I will update the checklist.
