### 2021-11-03 d6b1a00
Add training component switch
