### 2021-11-03 db1ca68
Debugging weight names
