### 2021-11-17 82f0965
Enable loading using string path
