### 2021-11-17 51a2cbe
More debugging of weight names
