### 2021-11-17 e55e592
More debugging statements
