### 2021-11-17 bade67d
Hardcode freezing logic
