### 2021-11-18 dba8c18
Fallback KL weight to current one on recompile
