### 2021-11-19 3dbec5d
Revert to flexible freezing + add checks
