### 2021-11-22 0435d66
Refactoring + change default training batch size
