### 2021-11-22 5849b66
Add weights duplication test
