### 2021-11-22 c5add8f
Boilerplate model benchmarking code
