+++ benchmarking/main.py
@@ -1,0 +1,372 @@
+"""Train and evaluate a MEGNetProbModel for formation energy prediction on the MP-2019 dataset.
+
+<https://figshare.com/articles/dataset/Graphs_of_Materials_Project_20190401/8097992>
+
+"""
+from abc import ABC, abstractmethod
+from argparse import ArgumentError, ArgumentParser
+from datetime import datetime
+from os import mkdir
+from pathlib import Path
+from typing import Iterable, List, NamedTuple, Optional, Tuple
+
+import pandas as pd
+from megnet.models import MEGNetModel
+from tensorflow.python.keras.callbacks import ModelCheckpoint, TensorBoard
+from unlocknn import MEGNetProbModel
+from unlocknn.megnet_utils import MEGNetGraph, Targets, default_megnet_config
+from unlocknn.metrics import MAE, MSE, evaluate_uq_metrics
+
+
+def eval_meg_model(
+    meg_model: MEGNetModel, graphs: Iterable[MEGNetGraph], targets: Iterable[float]
+) -> Tuple[float, float]:
+    """Evaluate the MSE and MAE of a given MEGNetModel."""
+    predictions = meg_model.predict_graphs(graphs)
+    return MSE(predictions, None, targets), MAE(predictions, None, targets)
+
+
+def train_megnet(
+    model: MEGNetModel,
+    num_epochs: int,
+    train_input,
+    train_targets,
+    val_input,
+    val_targets,
+    checkpoint_path: Path,
+    verbose: int = 2,
+    callbacks=[],
+):
+    """Train a MEGNetModel."""
+    ckpt_callback = ModelCheckpoint(
+        checkpoint_path, save_best_only=True, save_weights_only=True
+    )
+    callbacks.append(ckpt_callback)
+    model.train_from_graphs(
+        train_graphs=train_input,
+        train_targets=train_targets,
+        validation_graphs=val_input,
+        validation_targets=val_targets,
+        epochs=num_epochs,
+        save_checkpoint=False,
+        automatic_correction=False,
+        callbacks=callbacks,
+        verbose=verbose,
+    )
+
+
+class Dataset(NamedTuple):
+    """Container for datasets split into train, validation and test subsets."""
+
+    train_input: List[MEGNetGraph]
+    train_targets: Targets
+    val_input: List[MEGNetGraph]
+    val_targets: Targets
+    test_input: List[MEGNetGraph]
+    test_targets: Targets
+
+
+class UnlockTrainer(ABC):
+    """Handler for training and benchmarking unlockNN models."""
+
+    def __init__(self) -> None:
+        """Initialize parameters."""
+        super().__init__()
+
+        # * Read command line arguments
+        self.init_cli_args()
+
+        # * Create necessary folders
+        self.create_dir_structure()
+
+        # * Initialize datasets, must be populated with `self.load_data`
+        self.data: Optional[Dataset] = None
+
+        # * Initialize model variables
+        self.meg_model: Optional[MEGNetModel] = None
+        self.prob_model: Optional[MEGNetProbModel] = None
+
+    def create_dir_structure(self):
+        """Create the project directory tree."""
+        root_models_dir = self.root_dir / "models"
+        self.megnet_models_dir = root_models_dir / "megnet"
+        self.prob_models_dir = root_models_dir / "prob"
+
+        root_logs_dir = self.root_dir / "logs"
+        self.megnet_logs_dir = root_logs_dir / "megnet"
+        self.prob_logs_dir = root_logs_dir / "prob"
+
+        for dirname in [
+            root_models_dir,
+            self.megnet_models_dir,
+            self.prob_models_dir,
+            root_logs_dir,
+            self.megnet_logs_dir,
+            self.prob_logs_dir,
+        ]:
+            if not dirname.exists():
[…]
