+++ benchmarking/matbench/perovskites/benchmark.py
@@ -3,8 +3,8 @@ from ..matbench import MatbenchTrainer
 
 if __name__ == "__main__":
     trainer = MatbenchTrainer(
-        "matbench_phonons",
-        "https://ml.materialsproject.org/projects/matbench_phonons.json.gz",
-        "last phdos peak",
+        "matbench_perovskites",
+        "https://ml.materialsproject.org/projects/matbench_perovskites.json.gz",
+        "e_form",
     )
     trainer.execute()
