### 2021-11-22 0ffac10
Fix benchmark scripts
