+++ benchmarking/matbench/perovskites/benchmark.py
@@ -1,10 +1,13 @@
 """Benchmarking script for the matbench phonons dataset."""
-from ..matbench import MatbenchTrainer
+from pathlib import Path
+
+from .. import matbench_utils
 
 if __name__ == "__main__":
-    trainer = MatbenchTrainer(
+    trainer = matbench_utils.MatbenchTrainer(
         "matbench_perovskites",
         "https://ml.materialsproject.org/projects/matbench_perovskites.json.gz",
         "e_form",
+        root_dir=Path(__file__).parent,
     )
     trainer.execute()
