### 2021-11-23 0e6202e
Update default megnet metric
