### 2021-11-23 921d7ee
Fix series types
