### 2021-11-25 #3700 @a-ws-m
While I was running the benchmarks I discovered a compatibility issue with the earlier version of TensorFlow Probability I was using. This means I've had to train the models on my local machine, as the cluster's version of CUDA is incompatible with the newer version. The framework is all in place to train the models and I should be able to transfer it over to the new cluster I have just got access to today -- I'll keep you posted.
