### 2021-12-02 9706f15
Replace checkpoint loading attr with CLI arg
