### 2021-12-13 f25ab6a
Make root_dir a command line argument
