+++ .kanbn/index.md
@@ -1,0 +1,22 @@
+---
+startedColumns:
+  - 'In Progress'
+completedColumns:
+  - Done
+---
+
+# unlockNN
+
+## Backlog
+
+## Todo
+
+- [train-matbench-megnet-models](tasks/train-matbench-megnet-models.md)
+- [train-matbench-fine-tuning](tasks/train-matbench-fine-tuning.md)
+- [train-matbench-variation-point-locations](tasks/train-matbench-variation-point-locations.md)
+
+## In Progress
+
+- [get-unlock-nn-envrionment-working-on-ucl-rcs](tasks/get-unlock-nn-envrionment-working-on-ucl-rcs.md)
+
+## Done
