+++ .kanbn/tasks/train-matbench-variation-point-locations.md
@@ -1,0 +1,15 @@
+---
+created: 2021-12-13T21:06:38.667Z
+updated: 2021-12-13T21:08:45.915Z
+assigned: ""
+progress: 0
+tags: []
+---
+
+# Train matbench: variation point locations
+
+Freeze neural network layers and train the inducing point locations + kernel parameters (VGP).
+
+## Relations
+
+- [blocker train-matbench-megnet-models](train-matbench-megnet-models.md)
