### 2021-12-18 0c6415a
Fix required fold for data only mode
