### 2022-01-19 74ce5d8
Fix model dirname method
