### 2022-02-11 59ef78b
Fix ignore_ckpt CLI arg
