### 2022-02-11 c92881d
Fix numpy array inputs to `predict`
