### 2022-02-12 c3a0da4
Perovskites benchmark results + vis
