+++ .kanbn/tasks/train-matbench-variation-point-locations.md
@@ -1,9 +1,10 @@
 ---
 created: 2021-12-13T21:06:38.667Z
-updated: 2021-12-13T21:08:45.915Z
+updated: 2022-03-02T12:49:45.163Z
 assigned: ""
 progress: 0
 tags: []
+completed: 2022-03-02T12:49:45.163Z
 ---
 
 # Train matbench: variation point locations
