+++ .kanbn/tasks/train-matbench-megnet-models.md
@@ -1,9 +1,11 @@
 ---
 created: 2021-12-13T21:04:05.706Z
-updated: 2021-12-13T21:05:44.901Z
+updated: 2022-03-02T12:49:38.580Z
 assigned: ""
 progress: 0
 tags: []
+started: 2022-03-02T12:49:34.797Z
+completed: 2022-03-02T12:49:38.580Z
 ---
 
 # Train matbench: MEGNet models
