+++ .kanbn/tasks/correct-megnet-evaluation-metrics.md
@@ -1,0 +1,12 @@
+---
+created: 2022-03-02T12:50:28.498Z
+updated: 2022-03-02T12:50:28.495Z
+assigned: ""
+progress: 0
+tags: []
+started: 2022-03-02T12:50:28.498Z
+---
+
+# Correct MEGNet evaluation metrics
+
+Fix/check that MEGNet model metrics are evaluating correctly, given the difference in TensorFlow version.
