+++ benchmarking/visualise.py
@@ -1,9 +1,9 @@
 """Tools for visualising benchmarking results."""
 from collections import defaultdict
-import json
+from operator import itemgetter
 import re
 from pathlib import Path
-from typing import Dict, List, Tuple, Union
+from typing import Dict, List, Optional, Union
 
 import matplotlib.pyplot as plt
 import pandas as pd
@@ -12,17 +12,49 @@ import seaborn as sns
 Pathy = Union[str, Path]
 
 
-def read_results(results_dirs: List[Path], metric_fname: str = "metrics.csv") -> pd.DataFrame:
+sns.set_style()
+
+
+def read_prob_results(
+    results_dirs: Dict[int, List[Path]], metric_fname: str = "metrics.csv"
+) -> pd.DataFrame:
     """Read results from files into a DataFrame."""
-    data = defaultdict(list())
+    data = defaultdict(list)
+    for points, dirs in results_dirs.items():
+        for fold, path in enumerate(dirs):
+            metric_path = path / metric_fname
+            metric_series: pd.Series = pd.read_csv(metric_path, index_col=0).squeeze(
+                "columns"
+            )
+
+            for entry in metric_series.index:
+                data[entry].append(metric_series[entry])
+
+            data["fold"].append(fold)
+            data["points"].append(points)
+            data["Model name"].append(f"Unlock-{points}")
+
+    return pd.DataFrame(data)
+
+
+def read_base_results(
+    results_dirs: List[Path],
+    base_model_name: str = "Base",
+    metric_fname: str = "metrics.csv",
+):
+    """Read results from files into a DataFrame."""
+    data = defaultdict(list)
     for fold, path in enumerate(results_dirs):
         metric_path = path / metric_fname
-        metric_series: pd.Series = pd.read_csv(metric_path, index_col=0, squeeze=True)
+        metric_series: pd.Series = pd.read_csv(metric_path, index_col=0).squeeze(
+            "columns"
+        )
 
         for entry in metric_series.index:
             data[entry].append(metric_series[entry])
 
         data["fold"].append(fold)
+        data["Model name"].append(base_model_name)
 
     return pd.DataFrame(data)
 
@@ -34,6 +66,7 @@ class ResultsLoader:
         prob_model_results: Pathy,
         num_folds: int,
         model_name_pattern: str = r"\D*(?:-(?P<points>\d+)-)?(?P<fold>\d+)$",
+        base_model_name: str = "Base",
         metric_fname: str = "metrics.csv",
     ) -> None:
         """Initialize log paths."""
@@ -41,30 +74,108 @@ class ResultsLoader:
         self.prob_model_results = Path(prob_model_results)
         self.model_name_pattern = model_name_pattern
         self.num_folds = num_folds
+        self.base_model_name = base_model_name
         self.metric_fname = metric_fname
 
-    def results_directories(self, base_model: bool = False) -> Union[List[Path], Tuple[List[Path], List[int]]]:
+    def prob_results_directories(self) -> Dict[int, List[Path]]:
         """Get the results directories for the model.
-        
-        Args:
-            base_model: If True, get the results for the unmodified model.
-                Default (False) gives results for the probabilistic model.
-        
-        Returns:
-            A list of paths, ordered by fold index.
 
-            A list of inducing index points, only if ``base_model = False``.
-        
+        Returns:
+            A dictionary of ``{num_inducing_points: [path]}`` with the ``path``s
+            ordered by fold index.
+
         """
-        model_dirs = {}
-        results_dir = self.base_model_results if base_model else self.prob_model_results
-        for direct in results_dir.iterdir():
+        model_dirs = defaultdict(list)
+        for direct in self.prob_model_results.iterdir():
+            if not direct.is_dir():
+                continue
+
+            match = re.match(self.model_name_pattern, direct.name)
+            if match:
[…]
