### 2022-03-17 221d7ca
Fix MEGNet metric evaluation
