### 2022-05-03 61b90dd
Remove obsolete kanban
