+++ .kanbn/tasks/get-unlock-nn-envrionment-working-on-ucl-rcs.md
@@ -1,19 +1,0 @@
----
-created: 2021-12-13T21:05:23.752Z
-updated: 2022-03-02T12:49:36.614Z
-assigned: ""
-progress: 0
-tags: []
-started: 2021-12-13T00:00:00.000Z
-completed: 2022-03-02T12:49:36.615Z
----
-
-# Get unlockNN envrionment working on UCL RCS
-
-Currently there is an import error due to a missing C library:
-`ImportError: /home/ucecwsm/.conda/envs/unlockNN/lib/python3.8/site-packages/pymatgen/util/coord_cython.cpython-38-x86_64-linux-gnu.so: undefined symbol: __intel_sse2_strchr`
-
-## Sub-tasks
-
-- [x] Get in touch with RCS support
-- [ ] Action their recommendations
