+++ .kanbn/index.md
@@ -1,23 +1,0 @@
----
-startedColumns:
-  - 'In Progress'
-completedColumns:
-  - Done
----
-
-# unlockNN
-
-## Backlog
-
-## Todo
-
-## In Progress
-
-- [correct-megnet-evaluation-metrics](tasks/correct-megnet-evaluation-metrics.md)
-
-## Done
-
-- [get-unlock-nn-envrionment-working-on-ucl-rcs](tasks/get-unlock-nn-envrionment-working-on-ucl-rcs.md)
-- [train-matbench-variation-point-locations](tasks/train-matbench-variation-point-locations.md)
-- [train-matbench-fine-tuning](tasks/train-matbench-fine-tuning.md)
-- [train-matbench-megnet-models](tasks/train-matbench-megnet-models.md)
