+++ .kanbn/tasks/train-matbench-megnet-models.md
@@ -1,17 +1,0 @@
----
-created: 2021-12-13T21:04:05.706Z
-updated: 2022-03-02T12:49:38.580Z
-assigned: ""
-progress: 0
-tags: []
-started: 2022-03-02T12:49:34.797Z
-completed: 2022-03-02T12:49:38.580Z
----
-
-# Train matbench: MEGNet models
-
-Train the baseline MEGNet models on the matbench tasks.
-
-## Relations
-
-- [blocker get-unlock-nn-envrionment-working-on-ucl-rcs](get-unlock-nn-envrionment-working-on-ucl-rcs.md)
