+++ .kanbn/tasks/train-matbench-variation-point-locations.md
@@ -1,16 +1,0 @@
----
-created: 2021-12-13T21:06:38.667Z
-updated: 2022-03-02T12:49:45.163Z
-assigned: ""
-progress: 0
-tags: []
-completed: 2022-03-02T12:49:45.163Z
----
-
-# Train matbench: variation point locations
-
-Freeze neural network layers and train the inducing point locations + kernel parameters (VGP).
-
-## Relations
-
-- [blocker train-matbench-megnet-models](train-matbench-megnet-models.md)
