### 2022-05-31 #3700 @osorensen
Thanks @a-ws-m. Regarding @TahiriNadia's comments, note that she has already completed her review checklist, so fixing those issues should not be blocking. However, if you want, you could create those same issues in the source repository yourself, and make sure to include a link to this review issue.

Looking forward to seeing the updates.
