+++ benchmarking/visualise.py
@@ -3,9 +3,10 @@ from collections import defaultdict
 from operator import itemgetter
 import re
 from pathlib import Path
-from typing import Dict, List, Optional, Union
+from typing import Dict, List, Optional, Tuple, Union
 
 import matplotlib.pyplot as plt
+import numpy as np
 import pandas as pd
 import seaborn as sns
 
@@ -13,6 +14,124 @@ Pathy = Union[str, Path]
 
 
 sns.set_style()
+
+
+def calc_pis(
+    residuals: np.ndarray, stddevs: np.ndarray
+) -> Tuple[np.ndarray, np.ndarray]:
+    """Calculate probability intervals for given set of residuals and standard deviations."""
+    norm_resids = residuals / stddevs
+    predicted_pi = np.linspace(0, 1, 100)
+    bounds = norm.ppf(predicted_pi)
+    observed_pi = np.array([np.count_nonzero(norm_resids <= bound) for bound in bounds])
+    observed_pi = observed_pi / norm_resids.size
+    return predicted_pi, observed_pi
+
+
+def parity_plot(
+    prob_df: pd.DataFrame,
+    fname: str,
+    target_name: str,
+    top_padding: float = 1.0,
+    bottom_padding: float = 0.5,
+):
+    """Plot the parity with 95% CI error bars for a dataset."""
+    PREDICTED_NAME = f"Predicted {target_name}"
+    OBSERVED_NAME = f"Observed {target_name}"
+
+    # MARKER_COLOUR = tuple(np.array([168, 190, 240]) / 255)
+
+    plot_df = pd.DataFrame(
+        {
+            PREDICTED_NAME: prob_df["Predicted value"],
+            OBSERVED_NAME: prob_df["True value"],
+            "CI": prob_df["Predicted standard deviation"] * 2,
+            "# Inducing index points": prob_df["# Inducing points"],
+        }
+    )
+    Y_LIMS = (
+        min(plot_df[PREDICTED_NAME].min(), plot_df[OBSERVED_NAME].min())
+        - bottom_padding,
+        max(plot_df[PREDICTED_NAME].max(), plot_df[OBSERVED_NAME].max()) + top_padding,
+    )
+    g = sns.FacetGrid(data=plot_df, height=7, col="# Inducing index points", margin_titles=True)
+    g.map_dataframe(
+        plt.errorbar,
+        OBSERVED_NAME,
+        PREDICTED_NAME,
+        "CI",
+        marker="o",
+        # color=MARKER_COLOUR,
+        linestyle="",
+        alpha=0.7,
+        markeredgewidth=1,
+        markeredgecolor="black",
+    )
+
+    g.map(
+        sns.lineplot,
+        x=Y_LIMS,
+        y=Y_LIMS,
+        label="Ideal",
+        color="black",
+        linestyle="--",
+        marker="",
+    )
+    g.set(xlim=Y_LIMS, ylim=Y_LIMS, aspect="equal")
+    g.set_titles(col_template="{col_name} points")
+    plt.savefig(fname, transparent=True, bbox_inches="tight")
+
+
+def plot_calibration(prob_df: pd.DataFrame, fname):
+    """Plot a calibration curve for a given dataset."""
+    PREDICTED_NAME = "Predicted cumulative distribution"
+    OBSERVED_NAME = "Observed cumulative distribution"
+
+    # LINE_COLOUR = tuple(np.array([0, 40, 85]) / 255)
+    # FILL_COLOUR = tuple(np.array([203, 216, 246]) / 255)
+
+    data = None
+    for num_inducing_points, subdf in prob_df.groupby("# Inducing points"):
+        predicted_pi, observed_pi = calc_pis(
[…]
