### 2022-06-30 b78256b
Get average metrics and show legend
