+++ benchmarking/matbench/perovskites/plot_results.py
@@ -8,6 +8,7 @@ if __name__ == "__main__":
     meg_res_dir = result_dir / "megnet"
     prob_res_dir = result_dir / "prob"
     rl = visualise.ResultsLoader(meg_res_dir, prob_res_dir, 5, base_model_name="MEGNet")
+    rl.get_average_metrics()
     rl.plot_metric(
         "perovskites-mae-change.png",
         title="Matbench Perovskites: Adding UnlockNN",
@@ -30,7 +31,9 @@ if __name__ == "__main__":
         "perovskites-parity.png",
         target_name="Formation energy / eV per atom",
         num_scatter=150,
+        title="Matbench Perovskites: Probabilistic parity plots"
     )
     rl.calibration_plot(
-        "perovskites-calibration.png"
+        "perovskites-calibration.png",
+        "Matbench Perovskites: Calibration curves"
     )
