### 2022-06-30 45ec0e2
Finalise benchmarks info JOSS Review #19
