### 2022-06-30 a-ws-m/unlockNN#19 @a-ws-m
The benchmarks are complete and I've reported the results for a few models with different hyperparameters in the appropriate [results directory](https://github.com/a-ws-m/unlockNN/tree/joss-review/benchmarking/matbench/results).
