### 2022-06-30 #3700 @a-ws-m
Thanks @Het-Shah! There are several non-blocking issues on the repository that I addressed a while ago, I'll ask @pmeier to follow up on those separately. I think the only remaining task is RE: substantial scholarly effort.

UnlockNN is specifically tailored to those working within chemistry, hence the focus on applying the uncertainty quantification to ``MEGNet``, which is a successful model for materials property prediction. The bindings needed to make the modifications in this context are quite substantial and required a lot of trial-and-error. Furthermore, for the intended audience, including experimental (rather than computational) researchers, the skills to implement these are less common than for a full-time developer. To that end, the examples are designed to make the package more accessible to them, and the benchmarks are specifically for materials property prediction tasks.

The API component is not limited to applications in chemistry; in principle it should work with any Tensorflow model. Its function is, as [in this comment](https://github.com/openjournals/joss-reviews/issues/3700#issuecomment-932956088), to replace one or more layers of a neural network. I think this is still not trivial due to some quirks of the software packages involved and the method by which the resulting network must be trained, which are accounted for in unlockNN. Although this is now achieved in relatively few lines of code, the [original release](https://github.com/a-ws-m/unlockNN/tree/f90784347cdea83259382fe12885bdf56a72e17f) had ~2400 LoC, which was mostly rewritten; a lot of effort has been invested into streamlining the code.
