### 2022-06-30 #3700 @a-ws-m
Ah I realise I'd already written a response to the issue about substantial scholarly effort in the thread [here](https://github.com/openjournals/joss-reviews/issues/3700#issuecomment-950791488); I think it's complementary to my last comment. Please let me know if I can provide any more details to address this last task @pmeier.
