### 2022-07-01 #3700 @pmeier
@osorensen This review dragged on for a lot longer than I anticipated. I'm currently on parental leave and won't have time to do substantial work for this until approximately mid August.

Seeing that @Het-Shah now accepted the submission with a "Substantial scholarly effort" while he shared similar concerns as I did in the beginning, I would also be fine to just accept it to move forward. Thoughts?
