### 2022-07-01 8ac7766
Minor edit for references
References to papers published in *npj Computational Materials* appear as *Npj Computational Materials* in the reference list. This should fix the issue. Otherwise the paper reads very well, so I have no further comments.

https://github.com/openjournals/joss-reviews/issues/3700#issuecomment-1172521759
