### 2022-07-05 1f7735c
Bump version + docs build fix
