+++ setup.py
@@ -11,7 +11,7 @@ README = (HERE / "README.md").read_text()
 # This call to setup() does all the work
 setup(
     name="unlockNN",
-    version="2.0.2",
+    version="2.0.3",
     description="Uncertainty quantification for neural network models of chemical systems.",
     long_description=README,
     long_description_content_type="text/markdown",
