+++ dev_environment.yml
@@ -25,7 +25,7 @@ dependencies:
       - megnet
       - requests
       - sphinx
-      - tensorflow
-      - tensorflow-probability
+      - tensorflow==2.7.0
+      - tensorflow-probability==0.15.0
       - tox
       - twine
