### 2022-07-05 #3700 @osorensen
> Thanks @a-ws-m. Once you've merged the changes into the master branch, could you please
>
> * [ ]  Make a tagged release of your software, and list the version tag of the archived version here.
> * [ ]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
> * [ ]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
> * [ ]  Please list the DOI of the archived version here.
>
> Meanwhile I will read through the manuscript once more, and let you know if I have any suggested changes.

@a-ws-m please let me know when this is done, and I can move forward with accepting the submission.
