### 2022-07-05 #3700 @a-ws-m
Hi, I've made the release: it's [version 2.0.3](https://github.com/a-ws-m/unlockNN/releases/tag/v2.0.3).
The Zenodo DOI is 10.5281/zenodo.6799685.
