### 2022-07-05 2f7bce8
Journal reference to arXiv preprint
Related to [JOSS submission](https://github.com/openjournals/joss-reviews/issues/3700).

I think this might be the correct reference for Tran et al. (2020)? If not, let's keep the arXiv reference as is.
